requirements.txt
streamlit>=1.28.0
streamlit-lottie
langchain-google-genai
langchain-core
langchain-community
pydantic
python-dotenv
google-generativeai
geopy
googletrans==4.0.0rc1
requests
Pillow
plotly
pandas
bcrypt
pygame
streamlit-modal
streamlit_js_eval
scikit-learn
joblib
numpy
PyJWT
statsmodels
prophet
timezonefinder
meteostat
textblob
transformers
torch
streamlit-drawable-canvas
authlib
google-auth
google-auth-oauthlib
google-auth-httplib2
fpdf
nltk
